Added Placeholders and Templates
I've added a very basic setup for the start of how I plan to design the website. I want to make it overly simplistic since it's based off the idea or scheme of a terminal. (Obviously, there will be things like images and videos that'll help with it but it's just the theme I want to go for.)

Considering in making important pieces of text green and somehow adding borders around my dividers. (header, nav, main, footer).