Comments, Styles and Button
Added comments explaining a few bits of the style I decided to go with the site.

Making a very short and easy additional site for a button. Mostly to show my work and because I'm not too certain how to show my button without it being useless. So I figured I'd kill two birds with one stone by adding a link to the second site and a button to the site.

I think I have an idea on what to do for the site but I will let you know the next time I start to work on it heavily.