Removed The Quiz
Yeah, I decided last second to redo some of the content and make it more just about the hobby more than a learning experience. It kind of sucks that I went through quite a bit of effort for something that didn't happen but at very least I was able to use most of the main file stuff and it might even be a good idea for the future.

Made the style a bit greener but I think I may have made it too green. I'm going to try to cut back on the green next time and add a few more images. Otherwise, flex is finally serving me well.

Speaking of media, I also added some music and plan on making the feedback section.... despite not really having anything in it at the moment. I plan to fix it in a few moments though.