I Keep Lying - Basically, I am Editing Again
I get to work a bit more on the assignment due to being able to work on it during this lab. With that in mind, I have changed the color of the 'a' tag text from blue to yellow so it looks a bit nicer. I have fixed the style of the video so that it is one half text and another half video. So it does look a bit better.